// apps/api/src/game/dto-interface/matchInfo.dto.ts
export class matchInfoDto {
  //  general
  name: string = 'normal';
  //  interface
  barLarge: number = 10;
  //  game
  gamexsize: number = 780;
  gameysize: number = 380;
  barDist: number = 20;
  barSpeed: number = 2;
  ballSpeed: number = 4;
  ballDeb: number = 150;
  ballSize: number = 10;
  barSize: number = 100;
  itemSize: number = 10;
  // not in setting
  ballx: number = 100;
  bally: number = 100;
  oneBary: number = 10;
  twoBary: number = 10;
  ballDirx: number = -1;
  ballDiry: number = -0.4;
  oneScore: number = 0;
  twoScore: number = 0;
  itemx: number = 40;
  itemy: number = 40;
  borderSize: number = 10;
  midSquareSize: number = 10;
  menuSize: number = 90;
  ysize: number = 500;
  xsize: number = 800;
  gamey: number = 400;
  gamex: number = 780;
}